[LO] Fix poor statement scope tracking leading to missing locals
Self-calling functions wouldnt work because a critical scope would get lost after it was overwritten in executeStatlist. Fixed this by using the same scope approach as executeExpression, which we probably should've done from the start.

Oh and, thanks to the last few commits, the executor can now be parsed by the Parser, ran through the Executor, and then used to run Parsed code
AKA The executor can now (successfully) obfuscate and execute the executor (so its probably bug free! (mostly))